requirements.txt
catboost
fastapi
httpx
pandas
psycopg[binary]
python-dotenv
sqlalchemy
streamlit
uvicorn
